+++ src/compositor.rs
@@ -235,6 +235,7 @@ fn start_compositor2(
         wait_for_sync_obj: Rc::new(WaitForSyncObj::new(&ring, &engine)),
         explicit_sync_enabled: Cell::new(true),
         keyboard_state_ids: Default::default(),
+        security_context_acceptors: Default::default(),
     });
     state.tracker.register(ClientId::from_raw(0));
     create_dummy_output(&state);