+++ src/globals.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
         backend::Backend,
-        client::Client,
+        client::{Client, ClientCaps},
         ifs::{
             ext_foreign_toplevel_list_v1::ExtForeignToplevelListV1Global,
             ext_idle_notifier_v1::ExtIdleNotifierV1Global,
@@ -34,6 +34,7 @@ use {
             wp_cursor_shape_manager_v1::WpCursorShapeManagerV1Global,
             wp_fractional_scale_manager_v1::WpFractionalScaleManagerV1Global,
             wp_presentation::WpPresentationGlobal,
+            wp_security_context_manager_v1::WpSecurityContextManagerV1Global,
             wp_single_pixel_buffer_manager_v1::WpSinglePixelBufferManagerV1Global,
             wp_tearing_control_manager_v1::WpTearingControlManagerV1Global,
             wp_viewporter::WpViewporterGlobal,
@@ -113,8 +114,8 @@ pub trait Global: GlobalBase {
     fn singleton(&self) -> bool;
     fn version(&self) -> u32;
     fn break_loops(&self) {}
-    fn secure(&self) -> bool {
-        false
+    fn required_caps(&self) -> ClientCaps {
+        ClientCaps::none()
     }
     fn xwayland_only(&self) -> bool {
         false
@@ -184,6 +185,7 @@ impl Globals {
         add_singleton!(ZwpVirtualKeyboardManagerV1Global);
         add_singleton!(ZwpInputMethodManagerV2Global);
         add_singleton!(ZwpTextInputManagerV3Global);
+        add_singleton!(WpSecurityContextManagerV1Global);
     }
 
     pub fn add_backend_singletons(&self, backend: &Rc<dyn Backend>) {
@@ -215,7 +217,7 @@ impl Globals {
 
     fn insert(&self, state: &State, global: Rc<dyn Global>) {
         self.insert_no_broadcast_(&global);
-        self.broadcast(state, global.secure(), global.xwayland_only(), |r| {
+        self.broadcast(state, global.required_caps(), global.xwayland_only(), |r| {
             r.send_global(&global)
         });
     }
@@ -223,11 +225,13 @@ impl Globals {
     pub fn get(
         &self,
         name: GlobalName,
-        allow_secure: bool,
+        client_caps: ClientCaps,
         allow_xwayland_only: bool,
     ) -> Result<Rc<dyn Global>, GlobalsError> {
         let global = self.take(name, false)?;
-        if (global.secure() && !allow_secure) || (global.xwayland_only() && !allow_xwayland_only) {
+        if client_caps.not_contains(global.required_caps())
+            || (global.xwayland_only() && !allow_xwayland_only)
+        {
             return Err(GlobalsError::GlobalDoesNotExist(name));
         }
         Ok(global)
@@ -236,7 +240,7 @@ impl Globals {
     pub fn remove<T: WaylandGlobal>(&self, state: &State, global: &T) -> Result<(), GlobalsError> {
         let _global = self.take(global.name(), true)?;
         global.remove(self);
-        self.broadcast(state, global.secure(), global.xwayland_only(), |r| {
+        self.broadcast(state, global.required_caps(), global.xwayland_only(), |r| {
             r.send_global_remove(global.name())
         });
         Ok(())
@@ -247,14 +251,16 @@ impl Globals {
     }
 
     pub fn notify_all(&self, registry: &Rc<WlRegistry>) {
-        let secure = registry.client.secure;
+        let caps = registry.client.effective_caps;
         let xwayland = registry.client.is_xwayland;
         let globals = self.registry.lock();
         macro_rules! emit {
             ($singleton:expr) => {
                 for global in globals.values() {
                     if global.singleton() == $singleton {
-                        if (secure || !global.secure()) && (xwayland || !global.xwayland_only()) {
+                        if caps.contains(global.required_caps())
+                            && (xwayland || !global.xwayland_only())
+                        {
                             registry.send_global(global);
                         }
                     }
@@ -268,11 +274,11 @@ impl Globals {
     fn broadcast<F: Fn(&Rc<WlRegistry>)>(
         &self,
         state: &State,
-        secure: bool,
+        required_caps: ClientCaps,
         xwayland_only: bool,
         f: F,
     ) {
-        state.clients.broadcast(secure, xwayland_only, |c| {
+        state.clients.broadcast(required_caps, xwayland_only, |c| {
             let registries = c.lock_registries();
             for registry in registries.values() {
                 f(registry);